+++ src/components/biddingPage/status/average.tsx
