+++ src/scss/components/biddingPage/_index.scss
@@ -1,3 +1,5 @@
+@import './status/';
+
 .bidding-container {
   display: flex;
   flex-wrap: wrap;
