+++ src/scss/components/biddingPage/status/_index.scss
@@ -1,0 +1,22 @@
+.bidding-status-container {
+  flex-basis: 100%;
+  height: 100%;
+  display: flex;
+  flex-wrap: wrap;
+}
+
+@import './participants.scss';
+@import './average.scss';
+
+.bidding-status-participants-container-out {
+  flex-basis: 32%;
+  margin-left: 5%;
+  margin-right: 3%;
+  height: 100%;
+}
+
+.bidding-status-average-container-out {
+  flex-basis: 57%;
+  margin-right: 3%;
+  height: 100%;
+}
