+++ src/scss/components/biddingPage/status/participants.scss
@@ -1,0 +1,44 @@
+.bidding-status-participants-container-in {
+  flex-basis: 100%;
+  height: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  background-color: white;
+  border-radius: 5px;
+  // border: #707070 solid 1px;
+}
+
+.bidding-status-participants-img {
+  flex-basis: 30%;
+  margin-left: 5%;
+  margin-right: 5%;
+  height: 40%;
+  margin-top: 25%;
+  // border: red solid 1px;
+}
+
+.bidding-status-participants-text-container {
+  flex-basis: 50%;
+  height: 100%;
+  display: flex;
+  flex-wrap: wrap;
+  color: #707070;
+}
+
+.bidding-status-participants-text-value {
+  flex-basis: 100%;
+  height: 25%;
+  margin-top: 50%;
+  text-align: center;
+  font-size: 36px;
+  // border: red solid 1px;
+}
+
+.bidding-status-participants-text-title {
+  flex-basis: 100%;
+  height: 20%;
+  margin-bottom: 20%;
+  font-size: 12px;
+  text-align: center;
+  // border: red solid 1px;
+}
